requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qdrant-client==1.8.0
cohere==4.57
asyncpg==0.29.0
sqlalchemy==2.0.23
openai==1.3.8
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.26.2